refactor: separate tracker and projects under datasets maincolumn (#2936)
