--- components/HCABioNetworks/Network/Atlas/components/Datasets/components/MainColumn/Table/tracker/types.ts
@@ -1,5 +1,0 @@
-import type { TrackerSourceDataset } from "../../../../../../../../../../@types/network";
-
-export interface Props {
-  data: TrackerSourceDataset[];
-}
