+++ components/HCABioNetworks/Network/Atlas/components/Datasets/components/MainColumn/tracker/components/Table/hook.ts
@@ -1,8 +1,8 @@
 import { SORT_DIRECTION } from "@databiosphere/findable-ui/lib/config/entities";
 import { Table, useReactTable } from "@tanstack/react-table";
-import type { TrackerSourceDataset } from "../../../../../../../../../../@types/network";
-import { CORE_OPTIONS } from "../../../../../../../../../common/Table/options/core/constants";
-import { SORTING_OPTIONS } from "../../../../../../../../../common/Table/options/sorting/constants";
+import type { TrackerSourceDataset } from "../../../../../../../../../../../@types/network";
+import { CORE_OPTIONS } from "../../../../../../../../../../common/Table/options/core/constants";
+import { SORTING_OPTIONS } from "../../../../../../../../../../common/Table/options/sorting/constants";
 import { COLUMNS } from "./columns";
 
 /**
