+++ components/HCABioNetworks/Network/Atlas/components/Datasets/components/MainColumn/tracker/components/Table/types.ts
@@ -1,0 +1,5 @@
+import type { TrackerSourceDataset } from "../../../../../../../../../../../@types/network";
+
+export interface Props {
+  data: TrackerSourceDataset[];
+}
